requirements.txt
anthropic
python-dotenv
openai
